New module "consent"
