+++ djangodenmark/settings/base.py
@@ -24,6 +24,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'membership',
+    'django_consent',
 ]
 
 MIDDLEWARE = [
